# server/internal/dto/user.go
package dto

import "github.com/bookpanda/messenger-clone/internal/model"

type UserUpdateRequest struct {
	Name              string `json:"name"`
	ProfilePictureURL string `json:"profilePictureUrl"`
}

type UserResponse struct {
	ID                uint   `json:"id"`
	Name              string `json:"name"`
	Email             string `json:"email"`
	ProfilePictureURL string `json:"profilePictureUrl"`
}

func ToUserResponse(user model.User) UserResponse {
	return UserResponse{
		ID:                user.ID,
		Name:              user.Name,
		Email:             user.Email,
		ProfilePictureURL: user.ProfilePictureURL,
	}
}

func ToUserResponseList(users []model.User) []UserResponse {
	userResponses := make([]UserResponse, len(users))
	for i, user := range users {
		userResponses[i] = ToUserResponse(user)
	}
	return userResponses
}